Update default for Template#render buffer argument
As of https://github.com/rails/rails/pull/34952, the fallback for nil
buffer was moved up to the method call's default argument. The `nil`
default here is passed to the original method (`old_render`), preventing
the default from being applied. This meant that any template going
through this monkey patch on Rails 6 blew up when it tried to call
methods on the buffer.